Remove unused & broken platform component
This is inherited from Moddable Platformer. It is not used in this game,
and it is also broken because this line of code:

    const SPRITE: Texture2D = preload("res://assets/tiles-a.png")

references an asset not present in this repo. This doesn't affect the
game at runtime, because this component is not used, but it does show a
big error in the editor every time the project is launched.