tileset: Remove patterns referencing nonexistent atlas source
Previously, clicking on either of the TileMapLayer nodes in main.tscn
would show this error in the Output tab, in red, dozens of times:

    ERROR: No TileSet atlas source with id 0.

The cause is that the TileSet defines a number of patterns using a tile
source that has been removed. I guess this is left-over from the TileSet
in Moddable Platformer which the current TileSet was adapted from.
(Upstream issue https://github.com/godotengine/godot/issues/89923 tracks
improving the editor behaviour here to make the error more obvious.)

Remove the unusuable patterns that caused this error.