Player: Remove all animations except idle
Previously, the player script would try to play a `walk` or `jump_down`
animation when the player is walking or falling (respectively). These
don't exist; so on every frame an error would be logged.

Remove this logic. Play the 'idle' animation automatically. Preserve
only the code to flip the player sprite when moving left or right.

Fixes https://github.com/endlessm/threadbare-origins-inkdrinker/issues/3